+++ src/test/php/net/xp_framework/unittest/io/streams/GzDecompressingInputStreamTest.class.php
@@ -37,13 +37,23 @@ class GzDecompressingInputStreamTest extends AbstractDecompressingInputStreamTes
     return gzencode($in, $level);
   }
 
-  #[@test]
-  public function with_original_filename() {
-    $decompressor= $this->newStream(new MemoryInputStream(
-      "\x1F\x8B\x08\x08\x82\x86\xE0T\x00\x03test.txt\x00\xF3H\xCD\xC9\xC9\x07\x00\x82\x89\xD1\xF7\x05\x00\x00\x00"
-    ));
+  /** @return var[][] */
+  private function dataWithFileName() {
+    return ["\x1F\x8B\x08\x08\x82\x86\xE0T\x00\x03test.txt\x00\xF3H\xCD\xC9\xC9\x07\x00\x82\x89\xD1\xF7\x05\x00\x00\x00"];
+  }
+
+  #[@test, @values('dataWithFileName')]
+  public function data_with_original_filename($data) {
+    $decompressor= $this->newStream(new MemoryInputStream($data));
     $chunk= $decompressor->read();
     $decompressor->close();
     $this->assertEquals('Hello', $chunk);
+  }
+
+  #[@test, @values('dataWithFileName')]
+  public function header_with_original_filename($data) {
+    $decompressor= $this->newStream(new MemoryInputStream($data));
+    $decompressor->close();
+    $this->assertEquals('test.txt', $decompressor->header()['filename']);
   }
 }